Create custering_cutomerdata
Given dataset contains 200 records and five columns, two of which describe the customer’s annual income and spending score. The latter is a value from 0 to 100. The higher the number, the more this customer has spent with the company in the past: 
Using k means clustering creates 6 clusters of customers based on their spending pattern. 
•	Visualize the same in a scatter plot with each cluster in a different color scheme. 
•	Display the cluster labels of each point.(print cluster indexes) 
•	Display the cluster centers.